Renamed rendering package to visuals
