+++ src/main/java/de/officeryoda/fhysics/engine/collision/CollisionSolver.kt
@@ -5,9 +5,9 @@ import de.officeryoda.fhysics.engine.FhysicsCore.EPSILON
 import de.officeryoda.fhysics.engine.math.Vector2
 import de.officeryoda.fhysics.engine.objects.FhysicsObject
 import de.officeryoda.fhysics.engine.util.times
-import de.officeryoda.fhysics.rendering.UIController.Companion.borderFrictionDynamic
-import de.officeryoda.fhysics.rendering.UIController.Companion.borderFrictionStatic
-import de.officeryoda.fhysics.rendering.UIController.Companion.borderRestitution
+import de.officeryoda.fhysics.visual.UIController.Companion.borderFrictionDynamic
+import de.officeryoda.fhysics.visual.UIController.Companion.borderFrictionStatic
+import de.officeryoda.fhysics.visual.UIController.Companion.borderRestitution
 import kotlin.math.abs
 import kotlin.math.sqrt
