+++ src/main/java/de/officeryoda/fhysics/engine/objects/Circle.kt
@@ -6,7 +6,7 @@ import de.officeryoda.fhysics.engine.collision.CollisionInfo
 import de.officeryoda.fhysics.engine.collision.ContactFinder
 import de.officeryoda.fhysics.engine.math.Projection
 import de.officeryoda.fhysics.engine.math.Vector2
-import de.officeryoda.fhysics.rendering.Renderer
+import de.officeryoda.fhysics.visual.Renderer
 
 /**
  * You know what a circle is.
